+++ client/src/pages/Dashboard.tsx
@@ -1,32 +1,9 @@
 import useLogin from "../hooks/data-fetching/useLogin";
-import { ContainedButton } from "../components/ui-components/StyledComponents";
-import useLogout from "../hooks/data-fetching/useLogout";
-import { sampleuserCredentials as userCredentials } from "../constants/constants";
 import useAuth from "../hooks/useAuth";
 
 const Dashboard = () => {
   const loginMutate = useLogin();
-  const logoutMutate = useLogout();
-  const { user, login, logout } = useAuth();
-
-  const handleLogin = () => {
-    loginMutate.mutate(userCredentials, {
-      onSuccess: (data) => {
-        login(data.user);
-      },
-    });
-  };
-
-  const handleLogout = () => {
-    logoutMutate.mutate(
-      loginMutate.data?.accessToken || "INVALID ACCESS TOKEN",
-      {
-        onSuccess: () => {
-          logout();
-        },
-      }
-    );
-  };
+  const { user } = useAuth();
 
   return (
     <div
@@ -38,19 +15,11 @@ const Dashboard = () => {
         flexDirection: "row",
       }}
     >
-      <ContainedButton onClick={handleLogin} disabled={loginMutate.isPending}>
-        {loginMutate.isPending ? "Logging in...." : "Login"}
-      </ContainedButton>
-      <ContainedButton onClick={handleLogout}>
-        {logoutMutate.isPending ? "Logging out...." : "Logout"}
-      </ContainedButton>
       {loginMutate.isPending && <div>....Loading User</div>}
       {loginMutate.isError && (
         <div>Encountered error: Please try after some time</div>
       )}
-      {user && (
-        <p style={{ textWrap: "wrap" }}>{JSON.stringify(loginMutate.data)}</p>
-      )}
+      {user && <p>{JSON.stringify(user)}</p>}
     </div>
   );
 };
